Refactor event attendees into sub app of events
Before adding more functionality to the event
attendee section it made sense to move the
attendee related code into it's own sub app
below the event app.
