+++ src/apps/events/attendees/controllers/index.js
@@ -1,0 +1,5 @@
+const { renderAttendees } = require('./list')
+
+module.exports = {
+  renderAttendees,
+}
