+++ src/apps/events/attendees/labels.js
@@ -1,0 +1,8 @@
+const attendeeLabels = {
+  company: 'Company',
+  job_title: 'Job title',
+  attended_date: 'Date attended',
+  service_delivery: 'Service delivery',
+}
+
+module.exports = { attendeeLabels }
