+++ src/apps/events/attendees/router.js
@@ -1,0 +1,7 @@
+const router = require('express').Router()
+
+const { renderAttendees } = require('./controllers')
+
+router.get('/', renderAttendees)
+
+module.exports = router
