+++ src/apps/events/attendees/views/list.njk
