+++ src/apps/events/repos.js
@@ -47,28 +47,10 @@ async function getActiveEvents (token, createdOn) {
   return eventsResponse.results
 }
 
-async function fetchEventAttendees (token, eventId, page = 1) {
-  const limit = 10
-  const offset = limit * (page - 1)
-
-  const eventAttendees = await authorisedRequest(token, {
-    url: `${config.apiRoot}/v3/interaction`,
-    qs: {
-      limit,
-      offset,
-      event_id: eventId,
-      sortby: 'contact__last_name,contact__first_name',
-    },
-  })
-
-  return eventAttendees
-}
-
 module.exports = {
   saveEvent,
   fetchEvent,
   getEvents,
   getAllEvents,
   getActiveEvents,
-  fetchEventAttendees,
 }
