+++ src/apps/interactions/labels.js
@@ -56,15 +56,7 @@ const filters = {
   sector_descends: 'Sector',
 }
 
-const attendeeLabels = {
-  company: 'Company',
-  job_title: 'Job title',
-  attended_date: 'Date attended',
-  service_delivery: 'Service delivery',
-}
-
 module.exports = {
-  attendeeLabels,
   interaction,
   serviceDelivery,
   policyFeedback,
