+++ src/apps/interactions/transformers/index.js
@@ -3,7 +3,6 @@ const transformInteractionToListItem = require('./interaction-to-list-item')
 const transformInteractionFormBodyToApiRequest = require('./interaction-form-body-to-api')
 const transformInteractionResponseToViewRecord = require('./interaction-response-to-view')
 const transformInteractionListItemToHaveUrlPrefix = require('./interaction-list-item-to-have-url')
-const transformServiceDeliveryToAttendeeListItem = require('./service-delivery-to-attendee-list-item')
 
 module.exports = {
   transformInteractionResponseToForm,
@@ -11,5 +10,4 @@ module.exports = {
   transformInteractionFormBodyToApiRequest,
   transformInteractionResponseToViewRecord,
   transformInteractionListItemToHaveUrlPrefix,
-  transformServiceDeliveryToAttendeeListItem,
 }
